+++ transform/src/types.ts
@@ -64,6 +64,7 @@ export class Schema {
   public needsLink: string | null = null;
   public byteSize: number = 0;
   public deps: Schema[] = [];
+  public customJsonKind: string = "any";
   private _custom: boolean = false;
 
   set custom(value: boolean) {